Add book button and move files to relevant folders
relates #12
