+++ App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { TabNavigator } from 'react-navigation';
-import Appointments from './components/Appointments';
+import Appointments from './components/appointments/Appointments';
 import Portfolio from './components/portfolio/Portfolio';
 import Book from './components/portfolio/Book';
-import Stats from './components/Stats';
+import Stats from './components/stats/Stats';
 
 const App = TabNavigator(
   {
